+++ client/src/modules/BestWork/data/tags.tsx
@@ -1,0 +1,12 @@
+type tagsType = {
+  [x: string]: string;
+};
+
+export const tags: tagsType = {
+  html: 'HTML',
+  css: 'CSS',
+  js: 'JavaScript',
+  react: 'React',
+  angular: 'Angular',
+  node: 'NodeJS',
+};
